backend/requirements.txt
fastapi
uvicorn
jinja2
python-multipart
azure-storage-blob
azure-identity
azure-keyvault-secrets